Update pokemon.java
There's currently one major bugs and a few things that need to be refactored. The biggest bug is that the values of life for "p_health" and "c_health" are not shared among the classes, meaning that it's impossible to win a battle right now. Will fix soon. Other than that there is just some simple refactoring that needs to be done to clean up the program.